refactor(analysis): carve the transform family out of the analysis_sql façade
Buffer and centroid move to analysis_sql/transform.py — the operations that
replace the source geometry rather than combining it with a second input.
Dissolve is conceptually the third member, but its statement is an aggregate
the materialize worker renders; only its NON_GROUPABLE_COLUMN_TYPES guard
reaches this package, and that stayed in shared.

The three BUFFER_SLICE constants did NOT come along, though render_geodesic_
buffer is their only caller. They are the antimeridian threshold set: the same
±180 and one-UTM-zone numbers render_dateline_safe reasons about, and the pair
has to be read together to see why the buffer gate tests >= at exactly the
width it slices at. Co-locating them with today's only consumer would split a
threshold from the helper that explains it.

The façade is now what it claims to be: a docstring, the re-exports, the
six-branch dispatcher, and __all__. 1255 lines -> 193, with the other 1366
across five family modules, none over 400. The package docstring's line count
is corrected to 1255 here too: that is what test_layering.py's ratchet
measured, and the PR should not carry two numbers for one file.

Byte-identity: all 94 cases pass, and the façade still exports all 35 symbols
the single file defined.
